+++ perflib/const.go
