+++ perflib/perflib.go
