+++ perflib/perflib_test.go
