+++ perflib/utf16.go
